Adição da pasta queries
